Refactor bandplan code
